\name{selectiveCI-package}
\alias{selectiveCI-package}
\alias{selectiveCI}
\docType{package}
\title{Confidence intervals for selected parameters}
\description{
Computes several kinds of confidence intervals for selected parameters. 
Currently implemented CIs are: Shortest Length, Modified Pratt, Quasi-Conventional and Benjamini-Yekutieli.
}
\details{
\tabular{ll}{
Package: \tab selectiveCI \cr
Type: \tab Package\cr
Version: \tab 0.30.0 \cr
Date: \tab 2013-08-19\cr
License: \tab MIT \cr
}
}

\author{Jonathan Rosenblatt and Asaf Weinstein
Maintainer: Jonathan Rosenblatt <john.ros.work@gmail.com>
}

\references{
[1] Weinstein, Asaf, William Fithian, and Yoav Benjamini. 2013. “Selection Adjusted Confidence Intervals With More Power to Determine the Sign.” Journal of the American Statistical Association 108 (501): 165–176. 

[2] Benjamini, Y., and D. Yekutieli. 2005. “False Discovery Rate-Adjusted Multiple Confidence Intervals for Selected Parameters.” Journal of the American Statistical Association 100 (469): 71–81.

}

\keyword{ package }
\keyword{univar}
\keyword{htest}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
